clean up documentation in api, add TODO
The TODO is to clean up the documentation for the getDataOutages function,
which takes no parameters but has documentation indicating it does. It may be
useful to save the information about getting outages for only one timestamp elswhere,
though.

This commit also deletes a "proof of concept" function from earlier in our project that
was not being used anywhere.